move Searchable to concerns
